Add garage button press checking
Users now have a log of when a garage button was pressed. Not just when
the app or its voice assistant hook is used.

Make the table pretty and centered

Make the log table look a little nicer with some centering of the text
and the table itself.

Fix Typo